Move node classes to submodule
Cleans up the `__init__.py` file while making imports easier to write.
