+++ tests/test_readme_example.py
@@ -34,7 +34,7 @@ def test_example(ebnf_path):
     #This will visit each node in the parse tree and print the line where its text begins
     DepthFirst(ast.root, lambda node: print(node.startLine))
 
-    from parse_ebnf import Comment
+    from parse_ebnf.nodes import Comment
 
     #Finds each comment in the file and prints its text content
     for child in ast.root.children:
